+++ libs/core/src/lib/adapters/create-noop-reaction.function.ts
@@ -3,6 +3,6 @@ export function createNoopReaction<State>(): NoopReaction<State> {
 }
 
 export type NoopReaction<State> = (s: State) => State;
-export interface WithNoopReaction<State> {
+export type WithNoopReaction<State> = {
   noop: NoopReaction<State>;
-}
+};